backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pandas==2.2.2
Flask-Migrate==4.0.5
Flask-CORS==4.0.0